requirements.txt
certifi==2017.11.5
chardet==3.0.4
coverage==4.4.2
Django==1.11.7
django-countries==5.0
django-filter==1.1.0
django-nose==1.4.5
djangorestframework==3.7.3
drf-compound-fields==1.0.0
idna==2.6
nose==1.3.7
psycopg2==2.7.3.2
pytz==2017.3
requests==2.18.4
requests-mock==1.3.0
six==1.11.0
urllib3==1.22
